pyproject.toml
[project]
name = "metadata-xml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.2",
    "markupsafe>=2.1.1",
    "metadata-xml",
    "pyyaml>=6.0.1",
    "validators==0.20.0",
    "xsd-download",
    "yattag==1.14.0",
]

[tool.uv]
dev-dependencies = [
    "lxml>=5.3.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "ruff>=0.11.3",
]

[tool.uv.sources]
xsd-download = { git = "https://github.com/cioos-siooc/XML-Schema-Downloader.git" }
metadata-xml = { workspace = true }

[tool.ruff]
lint.extend-select = ["I"]